requirements.txt
openai>=1.0.0
requests
fastapi
jinja2
python-dotenv
beautifulsoup4
uvicorn
python-multipart
